Add files via upload
This is a tool that can access the strength of a password based on a criteria that make a password strength such as the length of the password, presence of upper and lower case letters, numbers and special characters. It also provide a feedback to users on the password's strength.